backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
httpx==0.27.0
python-dotenv==1.0.1
